+++ cargo-crev/CHANGELOG.md
@@ -22,6 +22,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Make `crate {goto,dir,open,expand}` assume `-u` outside of an existing Rust project.
 - Introduce trust and package review "overrides" which allow overriding (ignoring) specific
    trust / package review
+- Add `cargo crev wot log` to help understand your WoT.
 
 ## [0.22.2](https://github.com/dpc/crev/compare/cargo-crev-v0.21.4...v0.22.2) - 2022-01-11